Fix accommodation creation (fixes #35)
The beds formset is only included in the accommodation form if the accommodation exists already. This leaded to an error when the backend expects the management form of the formset to be present in the response. As a fix, the management form is now included, even if the actual bed formset is not.