[sphinx-doc] Insert type hints right after parameter description
The original function 'process_django_models()' just inserted the new type hints at the very end of the docstring lines.
In case the parameter list gets separated, e.g. because we want to point out which parameters were inherited from a base model and which not, the type hints were not recognized.
The new solution is a bit over-engineered, but has the advantage that all type hints are inserted right after the parameter description, which means they are recognized even if the list is split up.
It also adds support for related fields and reverse relations.