Fix rendered diff on XLIFF import
The change to cached properties had an unwanted side effect on the XLIFF import confirmation page:
Since the existing translation was also used in a form instance to validate the imported translation, it got assigned the changed values of the form instance.
This means that the imported diff only showed the imcoming content, not the changes to the previous import.