Allow dynamic export of database objects
Adds a `dump` function that exports all objects of selected models
as json.
Also adds a new api endpoint `/v3/database` which uses this function
to create a data export.

To preserve user privacy, only whitelisted models get exported and
user account information will not be included in the export.