Don't prefetch related links
This optimization does not seem to be worth it, because
there are lots of links that have to be prefetched, but
only a small amount of urls where the links are actually used.
This reduces load time for münchen by at least 500 ms on my pc,
even with caching disabled.