Fix poi table layout
Remove the `table-fixed` class. Before pr 2895, we had a css rule that
set every table listing tables width to `inherit`. This caused
`table-fixed` to have no effect, because using `table-fixed` without
a specific size is an error and causes css to fallback to `table-auto`,
which is the default value.
Now, that every table has `w-full` though, this rule is valid and causes
the table to get too large. Fortunately, simply removing this rule fixes
the table layout.