Fix preact setter types
We were using the typ `StateUpdater<_>` for setters, but the type of
`StateUpdater<T>` just resolves to `T | ((prevState: T) => T)`, which
means that it is not guaranteed to be callable.

The preact documentation is not really helpful:
https://preactjs.com/guide/v10/typescript/#typing-hooks

But in the source of `useState`, the setter is typed as
`Dispatch<StateUpdater<T>>`, so I decide to use the same type.